Second Commit

+++ admin/includes/view_posts.php
@@ -5,28 +5,38 @@
             <th>Author</th>
             <th>Title</th>
             <th>Category</th>
-            <th>Status</th>
             <th>Image</th>
             <th>Comments</th>
-            <th>Tags</th>
             <th>Date</th>
+            <th>Edit</th>
         </tr>
     </thead>
     <tbody>
 
         <?php
-        $query = "SELECT * FROM posts";
+        if(isset($_SESSION['user_role'])) {
+            $user = $_SESSION['username'];
+            if($_SESSION['user_role'] == 'admin') {
+                $query = "SELECT * FROM posts";
+            } else {
+                $query = "SELECT * FROM posts WHERE post_author = '$user'";
+            }
+        }
         $posts = mysqli_query($connect, $query);
+        
+        if(mysqli_num_rows($posts) < 1) {
+            ?>
+        <blockquote class="blockquote"><strong>You haven't posted anything! </strong><a href="posts.php?source=add_post">Add new post</a></blockquote>
+        <?php
+        }
 
         while($row = mysqli_fetch_assoc($posts)) {
             $postID = $row['post_id'];
             $postAuthor = $row['post_author'];
             $postTitle = $row['post_title'];
             $postCategory = $row['post_cat_id'];
-            $postStatus = $row['post_status'];
             $postImage = $row['post_image'];
             $postComment = $row['post_comment_count'];
-            $postTag = $row['post_tags'];
             $postDate = $row['post_date'];
 
             echo "<tr>";
@@ -43,15 +53,11 @@
             
             echo "<td>$catTitle</td>";
             
-            
-            
-            echo "<td>$postStatus</td>";
             echo "<td><img width='100px' src='../images/$postImage' alt='image'></td>";
             echo "<td>$postComment</td>";
-            echo "<td>$postTag</td>";
             echo "<td>$postDate</td>";
             echo "<td><a href='posts.php?source=edit_post&p_id={$postID}'>Edit</a><br>";
-            echo "<a class='text-danger' href='posts.php?delete={$postID}'>Delete</a></td>";
+            echo "<a onclick=\"javascript: return confirm('Delete the post?'); \" class='text-danger' href='posts.php?delete={$postID}'>Delete</a></td>";
             echo "</tr>";
         }
         ?>
